Cleanup: cast pthread_self() return value to unsigned long
pthread_t can map to other things that unsigned long (e.g. pointer).
Cast it to unsigned long for debug printing and for debug delay random
value purposes.

+++ tests/test_urcu_hash.c
@@ -202,7 +202,8 @@ unsigned long test_compare(const void *key1, size_t key1_len,
 void *thr_count(void *arg)
 {
 	printf_verbose("thread_begin %s, thread id : %lx, tid %lu\n",
-			"counter", pthread_self(), (unsigned long)gettid());
+			"counter", (unsigned long) pthread_self(),
+			(unsigned long) gettid());
 
 	rcu_register_thread();
 
@@ -544,7 +545,8 @@ int main(int argc, char **argv)
 	printf_verbose("Number of hash chains: %lu.\n",
 		nr_hash_chains);
 	printf_verbose("thread %-6s, thread id : %lx, tid %lu\n",
-			"main", pthread_self(), (unsigned long)gettid());
+			"main", (unsigned long) pthread_self(),
+			(unsigned long) gettid());
 
 	tid_reader = malloc(sizeof(*tid_reader) * nr_readers);
 	if (!tid_reader) {
